Work around potential crash when invalid UTF-8 encountered
This switches from using `stricmp_utf8()` and `stricmp_utf8_ex()` to `pfc::stringCompareCaseInsensitive` and `pfc::stringCompareCaseInsensitiveEx` when sorting tracks to group them into nodes.

This is due to `stricmp_utf8()` and `stricmp_utf8_ex()` seemingly always returning -1 when strings containing invalid UTF-8 are compared if the two strings are identical up to the point of the invalid UTF-8.

This becomes problematic when sorting a number of such strings as it would imply a contradictory order for them, which can cause the sorting function to try to read outside of the bounds of the list of items being sorted (and hence cause a crash).

`pfc::stringCompareCaseInsensitive` and `pfc::stringCompareCaseInsensitiveEx` don't have this problem (though are slightly slower from tests).