Change selection when right-clicking on unselected item
Previously, right-clicking on an unselected item would highlight that item but not change the selection. That behaviour is now particularly odd when multiple items are selected.

This changes the behaviour to instead permanently change the selection when right-clicking on an unselected item. Autosend can be skipped by holding Shift while right-clicking.