Fix inability to drag multiple items
This resolves a problem where, when multiple items were selected, trying to drag them would only drag the item that was clicked on.

There were various changes that had to be made to click handling to accommodate this, so that the items are (correctly) deselected on mouse up rather than mouse down (when they aren't dragged).