Fix inconsistencies in middle-click behaviour
This resolves some inconsistencies when middle-clicking on the focused item.

Previously, when middle-clicking on a non-focused item, it selected that item and deselected others. However, when middle-clicking on the focused item, it did not make any changes to the selection. This meant it did not select the item if it wasn't selected, nor did it deselect items other than the focused one that were selected.

Whether middle-clicking on a selected item instead preserve the selection remains an open question. This, however, makes the current behaviour consistent.