Fix 32- and 64-bit configuration compatibility
This resolves a problem where the saved list of views wouldn't be read correctly by a 32-bit build of foobar2000 if foo_uie_albumlist.dll.cfg was written by a 64 build of foobar2000, and vice versa.

This was due to a `size_t` being incorrectly written in the serialised configuration.

On upgrade, the configuration is migrates to a new variable with a configuration format that is compatible between 32- and 64-bit versions of foobar2000. The old configuration value is also maintained to avoid data loss when downgrading to an older version of Album list panel.