Fix writing of bad config data when node state is missing
This fixes a problem where bad configuration data was written when the node state was missing (for example, after doing a search with no results).

In some cases, this resulted in a crash when the panel was next initialised with that bad saved configuration data.

The configuration reading code was also updated to correctly handle the case where bad data had been previously saved.