Don't focus filter on window creation
The panel attempted to focus the filter edit box when the panel window was created. This was incorrect, and caused the main window to be momentarily activated if foobar2000 was opened hidden or minimised.

This change stops the filter edit box from being focused on window creation. (It's still focused when opening the filter via the context menu.)