Don’t treat consecutive vertical bars as one in view formatting string
This changes a behaviour where consecutive | characters in the evaluated view title formatting string were treated as if they were one |, causing subsequent (child) nodes to be brought up a level.

Such consecutive | characters are no longer treated as a single |, making behaviour more consistent with the standalone Album List.

Such a scenario will now create a node with a label of `(blank)` (in the standalone Album List, the empty string is used).

This might affect existing views where the old behaviour was being relied on. | can be used in `$if` and similar title formatting expressions, and so it should still be possible to replicate the old behaviour.

Note that the behaviour of branching expressions (such as `%<artist>%`) is unchanged, and these still do not create a node for values that are empty strings.

Some of the code to process formatted titles was rewritten and modernised as part of this.