Improve handling of empty strings in branching expressions
Previously, branching expressions (such as `%<artist>%`) would not create nodes for empty strings (note that this is not the same as a field being missing).

This was even the case when using an expression like `%<artist>% a|%album%` (with the ARTIST field field set to an empty string), when logically that should evaluate to ` a`.

This changes the behaviour so that empty strings do not have special handling and behave consistently with non-branching expressions (such as `%artist%`).