Add Ctrl+Backspace and Ctrl+A support to edit controls
This enables edit control enhancements that add support for using Ctrl+Backspace to delete the previous word, and Ctrl+A for selecting all text (on older versions of Windows without native Ctrl+A support).

This applies to the Filter edit control and all edit controls in Album list panel preferences.