Use custom drag image on Windows 10 and newer
This updates the panel to use a custom drag image on Windows 10 and newer, in line with Columns UI 3.2.0.

This is to avoid the default white drag image introduced on Windows 11, and also to have a dark themed drag image.

There is no change on older versions of Windows.

Additionally, the track count in the drag image is now formatted using using the {fmt} library and the user’s locale, which should mean the number format configured in Windows is followed more closely.